Remove unused copies of validate_string_dict
