+++ IPython/qt/client.py
@@ -37,27 +37,10 @@ class QtHBChannel(SuperQObject, HBChannel):
 
 from IPython.core.release import kernel_protocol_version_info
 
-from IPython.utils.py3compat import string_types, iteritems
-
 major_protocol_version = kernel_protocol_version_info[0]
 
 class InvalidPortNumber(Exception):
     pass
-
-# some utilities to validate message structure, these might get moved elsewhere
-# if they prove to have more generic utility
-
-
-def validate_string_dict(dct):
-    """Validate that the input is a dict with string keys and values.
-
-    Raises ValueError if not."""
-    for k,v in iteritems(dct):
-        if not isinstance(k, string_types):
-            raise ValueError('key %r in dict must be a string' % k)
-        if not isinstance(v, string_types):
-            raise ValueError('value %r in dict must be a string' % v)
-
 
 
 class QtZMQSocketChannel(SuperQObject):
